implementation_2/requirements.txt
torch
numpy
pandas
matplotlib
scipy
gym
networkx
plotly
python-pptx
